scraper/requirements.txt
playwright==1.41.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
apscheduler==3.10.4
pyyaml==6.0.1
python-dotenv==1.0.0
loguru==0.7.2
prometheus-client==0.19.0
httpx==0.26.0
pydantic==2.5.3
pydantic-settings==2.1.0
fastapi==0.108.0
uvicorn[standard]==0.25.0